Refactor: makeAssistant 관련 fetch 로직 커스텀 훅으로 통합 (#46)
