+++ src/features/assistantList/types/assistantListTypes.ts
@@ -12,23 +12,8 @@ export type TAssistantProfile = {
   createdAt: string;
 };
 
-export type TAssistantMakeQnA = {
-  q1: string;
-  q2: string;
-  q3: string;
-  q4: string;
-  q5: string;
-  q6: string;
-};
-
 export type TNewAiTrigger = { isGenerateButton: boolean };
 
 export type AssistantProfileListDTO = SuccessDTO & {
   result: TAssistantProfile[];
-};
-
-export type MakeAssistantQuestionResult = SuccessDTO & {
-  result: TAssistantMakeQnA & {
-    createdAt: string;
-  };
 };
