+++ src/features/makeAssistant/types/makeAssistantTypes.ts
@@ -1,15 +1,6 @@
 import { SuccessDTO } from "@/src/types/commonTypes";
 
-export type TMakeAssistantAnswers = {
-  q1: string;
-  q2: string;
-  q3: string;
-  q4: string;
-  q5: string;
-  q6: string;
-};
-
-export type TAiProfile = {
+export type TAssistantProfile = {
   aiProfileId: number;
   profileName: string;
   imageUrl: string;
@@ -21,6 +12,21 @@ export type TAiProfile = {
   createdAt: string;
 };
 
+export type TAssistantMakeQnA = {
+  q1: string;
+  q2: string;
+  q3: string;
+  q4: string;
+  q5: string;
+  q6: string;
+};
+
 export type MakeAssistantDTO = SuccessDTO & {
-  result: TAiProfile;
+  result: TAssistantProfile;
+};
+
+export type MakeAssistantQuestionDTO = SuccessDTO & {
+  result: TAssistantMakeQnA & {
+    createdAt: string;
+  };
 };
